Extract test of project module flattening
